Scheduler delete from db every 5 min
