+++ shhh/config.py
@@ -22,9 +22,9 @@ class DefaultConfig:
     CELERY_BROKER_URL = "redis://localhost:6379"
     CELERY_RESULT_BACKEND = "redis://localhost:6379"
 
-    REDIS_HOST = "localhost"
-    REDIS_PORT = "6379"
-    REDIS_URL = "redis://localhost:6379"
+    # REDIS_HOST = "localhost"
+    # REDIS_PORT = "6379"
+    # REDIS_URL = "redis://localhost:6379"
 
 
 class ProductionConfig(DefaultConfig):
@@ -35,4 +35,4 @@ class ProductionConfig(DefaultConfig):
     CELERY_BROKER_URL = "redis://redis:6379"
     CELERY_RESULT_BACKEND = "redis://redis:6379"
 
-    REDIS_HOST = "redis"
+    # REDIS_HOST = "redis"
