+++ shhh/sql/delete_expired_links.sql
@@ -1,0 +1,3 @@
+DELETE
+FROM links
+WHERE date_expires <= now();
